changed from int to long on identificationnumber

+++ AirLineAPI/Model/Passenger.cs
@@ -13,7 +13,7 @@ namespace AirLineAPI.Model
         [Required]
         [StringLength(30, MinimumLength =2, ErrorMessage ="Name must be longer than 2 characters and shorter than 30")]
         public string Name { get; set; }
-        public int IdentificationNumber { get; set; }
+        public long IdentificationNumber { get; set; }
         public ICollection<PassengerTimeTable> PassengerTimeTables { get; set; }
     }
